+++ txnbuild/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this
 file.  This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [v7.1.0](https://github.com/stellar/go/releases/tag/horizonclient-v7.1.0) - 2021-06-01
 
 ### New features